--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,22 +1,22 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
 	"fmt"
+	"github.com/mailru/easyjson"
 	"io"
-	"io/ioutil"
 	"regexp"
 	"strings"
 )
 
 type User struct {
-	ID       int
-	Name     string
-	Username string
-	Email    string
-	Phone    string
-	Password string
-	Address  string
+	ID       int    `json:"Id"`
+	Name     string `json:"Name"`
+	Username string `json:"Username"`
+	Email    string `json:"Email"`
+	Phone    string `json:"Phone"`
+	Password string `json:"Password"`
+	Address  string `json:"Address"`
 }
 
 type DomainStat map[string]int
@@ -32,27 +32,35 @@ func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 type users [100_000]User
 
 func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
+	s := bufio.NewScanner(r)
+	i := 0
 
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
+	for s.Scan() {
 		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
+		if err = easyjson.Unmarshal([]byte(s.Text()), &user); err != nil {
 			return
 		}
 		result[i] = user
+		i++
 	}
+
+	if s.Err() != nil {
+		err = s.Err()
+	}
+
 	return
 }
 
 func countDomains(u users, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 
+	r, err := regexp.Compile("\\." + domain)
+	if err != nil {
+		return nil, err
+	}
+
 	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
+		matched := r.MatchString(user.Email)
 		if err != nil {
 			return nil, err
 		}