--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -3,73 +3,50 @@ package hw10programoptimization
 import (
 	"bufio"
 	"fmt"
-	"github.com/mailru/easyjson"
 	"io"
 	"regexp"
 	"strings"
+
+	"github.com/mailru/easyjson"
 )
 
 type User struct {
-	ID       int    `json:"Id"`
-	Name     string `json:"Name"`
-	Username string `json:"Username"`
-	Email    string `json:"Email"`
-	Phone    string `json:"Phone"`
-	Password string `json:"Password"`
-	Address  string `json:"Address"`
+	ID       int    `json:"id"`
+	Name     string `json:"name"`
+	Username string `json:"username"`
+	Email    string `json:"email"`
+	Phone    string `json:"phone"`
+	Password string `json:"password"`
+	Address  string `json:"address"`
 }
 
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
 	s := bufio.NewScanner(r)
-	i := 0
+	result := make(DomainStat)
+
+	re, err := regexp.Compile("\\." + domain)
+	if err != nil {
+		return nil, fmt.Errorf("error in regexp compiling: %w", err)
+	}
 
 	for s.Scan() {
 		var user User
-		if err = easyjson.Unmarshal([]byte(s.Text()), &user); err != nil {
-			return
-		}
-		result[i] = user
-		i++
-	}
-
-	if s.Err() != nil {
-		err = s.Err()
-	}
-
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
-	result := make(DomainStat)
-
-	r, err := regexp.Compile("\\." + domain)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, user := range u {
-		matched := r.MatchString(user.Email)
-		if err != nil {
-			return nil, err
+		if err = easyjson.Unmarshal(s.Bytes(), &user); err != nil {
+			return nil, fmt.Errorf("error in json unmarshalling: %w", err)
 		}
 
-		if matched {
+		if re.MatchString(user.Email) {
 			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
 			num++
 			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
 		}
 	}
+
+	if s.Err() != nil {
+		return nil, fmt.Errorf("error in scanning data: %w", s.Err())
+	}
+
 	return result, nil
 }