## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

-----
## Результаты оптимизации
### Основные изменения:
* Переход к EasyJSON для unmarshall
* Прекомпиляция RegEx
* Построчный Scanner для входных данных
* Объединение чтения и обработки в одном цикле (функции)
### benchstat
```
name             old time/op  new time/op  delta
GetDomainStat-8   1.15s ±75%   0.15s ± 3%  -87.12%  (p=0.000 n=10+9)
```

### Результаты референсного теста
#### Начальные показатели
```
=== RUN   TestGetDomainStat_Time_And_Memory
    stats_optimization_test.go:46: time used: 602.9571ms / 300ms
    stats_optimization_test.go:47: memory used: 292Mb / 30Mb
    stats_optimization_test.go:49: 
        	Error Trace:	/hw10_program_optimization/stats_optimization_test.go:49
        	Error:      	"602957100" is not less than "300000000"
        	Test:       	TestGetDomainStat_Time_And_Memory
        	Messages:   	the program is too slow
--- FAIL: TestGetDomainStat_Time_And_Memory (19.16s)
FAIL
FAIL	/otus-golang-hw/hw10_program_optimization	19.161s
FAIL
```
#### Финальные показатели
```
=== RUN   TestGetDomainStat_Time_And_Memory
    stats_optimization_test.go:46: time used: 146.7601ms / 300ms
    stats_optimization_test.go:47: memory used: 11Mb / 30Mb
--- PASS: TestGetDomainStat_Time_And_Memory (1.82s)
PASS
ok  	github.com/mrumpel/otus-golang-hw/hw10_program_optimization	1.827s
```