@agneum 2022-08-07
Думаю, что причина такого значения `p=1.000` в количестве запусков
Подробнее вопрос рассмотрен в этом issue: `https://github.com/golang/go/issues/23471#issue-289658870`

Предлагаю запустить с флагом -count, например,

```
$ go test -bench=. -benchmem -count=10 >  new.txt
```
Результат
```
$ benchstat old.txt new.txt 
name                  old time/op    new time/op    delta
GetDomainStatLong-6      677ms ± 5%     198ms ± 0%   -70.80%  (p=0.000 n=10+10)
GetDomainStatShort-6     224ms ± 2%       0ms ± 1%  -100.00%  (p=0.000 n=10+10)

name                  old alloc/op   new alloc/op   delta
GetDomainStatLong-6      322MB ± 0%      12MB ± 0%   -96.30%  (p=0.000 n=10+10)
GetDomainStatShort-6     211MB ± 0%       0MB ± 0%  -100.00%  (p=0.000 n=9+9)

name                  old allocs/op  new allocs/op  delta
GetDomainStatLong-6      3.15M ± 0%     0.65M ± 0%   -79.48%  (p=0.000 n=10+10)
GetDomainStatShort-6     2.10M ± 0%     0.00M ± 0%  -100.00%  (p=0.000 n=10+10)
```